[core] move classes into module
