+++ core/src/main/scala/net.kemitix.s3thorp.core/Counters.scala
@@ -1,0 +1,5 @@
+package net.kemitix.s3thorp.core
+
+final case class Counters(uploaded: Int = 0,
+                          deleted: Int = 0,
+                          copied: Int = 0)
