+++ core/src/main/scala/net.kemitix.s3thorp.core/KeyGenerator.scala
@@ -1,0 +1,17 @@
+package net.kemitix.s3thorp.core
+
+import java.io.File
+
+import net.kemitix.s3thorp.domain.RemoteKey
+
+object KeyGenerator {
+
+  def generateKey(source: File, prefix: RemoteKey)
+                 (file: File): RemoteKey = {
+    val otherPath = file.toPath.toAbsolutePath
+    val sourcePath = source.toPath
+    val relativePath = sourcePath.relativize(otherPath)
+    RemoteKey(s"${prefix.key}/$relativePath")
+  }
+
+}
