+++ src/test/resources/net/kemitix/s3thorp/awssdk/test-file-for-hash.txt
@@ -1,0 +1,1 @@
+This is a test file for checking the md5 hash generator.
