+++ src/test/resources/net/kemitix/s3thorp/awssdk/upload/root-file
@@ -1,0 +1,1 @@
+This file is in the root directory of the upload tree.
