+++ src/test/resources/net/kemitix/s3thorp/awssdk/upload/subdir/leaf-file
@@ -1,0 +1,1 @@
+This file is in the subdir folder within the upload tree.
