--- src/test/resources/net/kemitix/s3thorp/test-file-for-hash.txt
@@ -1,1 +1,0 @@
-This is a test file for checking the md5 hash generator.
