--- src/test/resources/net/kemitix/s3thorp/upload/root-file
@@ -1,1 +1,0 @@
-This file is in the root directory of the upload tree.
