--- src/test/resources/net/kemitix/s3thorp/upload/subdir/leaf-file
@@ -1,1 +1,0 @@
-This file is in the subdir folder within the upload tree.
