--- src/test/scala/net/kemitix/s3thorp/Resource.scala
@@ -1,14 +1,0 @@
-package net.kemitix.s3thorp
-
-import java.io.{File, FileNotFoundException}
-
-import scala.util.Try
-
-object Resource {
-
-  def apply(base: AnyRef,
-            name: String): File = {
-    Try(new File(base.getClass.getResource(name).getPath))
-      .getOrElse(throw new FileNotFoundException(name))
-  }
-}
