+++ src/test/scala/net/kemitix/s3thorp/awssdk/S3ClientMultiPartTransferManagerSuite.scala
@@ -7,17 +7,17 @@ import com.amazonaws.AmazonClientException
 import com.amazonaws.services.s3.model
 import com.amazonaws.services.s3.transfer.model.UploadResult
 import com.amazonaws.services.s3.transfer.{PauseResult, PersistableUpload, Transfer, TransferManager, TransferManagerBuilder, TransferProgress, Upload}
-import net.kemitix.s3thorp.KeyGenerator.generateKey
+import net.kemitix.s3thorp.core.KeyGenerator.generateKey
 import net.kemitix.s3thorp.aws.api.S3Action.UploadS3Action
 import net.kemitix.s3thorp.aws.api.UploadProgressListener
 import net.kemitix.s3thorp.domain.{Bucket, Config, LastModified, LocalFile, MD5Hash, RemoteKey}
-import net.kemitix.s3thorp.{KeyGenerator, MD5HashGenerator, Resource}
+import net.kemitix.s3thorp.core.{MD5HashGenerator, Resource, Sync}
 import org.scalatest.FunSpec
 
 class S3ClientMultiPartTransferManagerSuite
   extends FunSpec {
 
-  private val source = Resource(this, "..")
+  private val source = Resource(this, ".")
   private val prefix = RemoteKey("prefix")
   implicit private val config: Config = Config(Bucket("bucket"), prefix, source = source)
   implicit private val logInfo: Int => String => Unit = l => m => ()
